Description
---
This is a Base feature branch that all additional work on the implementation will be merged into, before being merged into development.

Motivation and Context
---
Removing the JS dependencies in favor of a pure rust implementation

Related
---
Toolchain upgrade:
https://github.com/tari-labs/docker/pull/23
https://github.com/tari-project/action-buildlibs/pull/10

How Has This Been Tested?
---
CI

Closes: #5050

+++ base_layer/core/src/base_node/sync/header_sync/synchronizer.rs
@@ -421,7 +421,7 @@ impl<'a, B: BlockchainBackend + 'static> HeaderSynchronizer<'a, B> {
                 },
             };
 
-            let steps_back = resp.fork_hash_index as u64 + offset as u64;
+            let steps_back = resp.fork_hash_index + offset as u64;
             return Ok((resp, block_hashes, steps_back));
         }
     }
